Fix OOM during 3B+ model quantization and e2e test failures
Three memory optimizations that make quantization viable on consumer hardware:

- Share TurboQuant rotation matrices across layers with the same dimensions,
  saving ~20GB for a 3B model (36 down_proj layers shared one (d,d) matrix
  instead of each allocating their own)
- Return int16 indices from LloydMax quantizer instead of int64 (4x reduction)
- Fix dtype mismatch in compressed forward path (float32 output vs float16 model)

Test fixes: load models in float16, split reference/quantized generation to
reduce peak memory, use lightweight name-only fixtures for weight quant tests.

+++ turboquant/lloyd_max.py
@@ -137,7 +137,7 @@ class LloydMaxQuantizer:
         """
         # Use searchsorted on boundaries for O(log n) quantization
         indices = torch.searchsorted(self.boundaries, x.contiguous())
-        return indices.clamp(0, self.n_levels - 1)
+        return indices.clamp(0, self.n_levels - 1).to(torch.int16)
 
     def dequantize(self, indices: torch.Tensor) -> torch.Tensor:
         """Map quantization indices back to centroid values.
